feat : swagger 작성

+++ src/main/java/com/yes255/yes255booksusersserver/presentation/controller/UserController.java
@@ -1,47 +1,52 @@
 package com.yes255.yes255booksusersserver.presentation.controller;
 
 import com.yes255.yes255booksusersserver.application.service.UserService;
-import com.yes255.yes255booksusersserver.persistance.repository.JpaCustomerRepository;
-import com.yes255.yes255booksusersserver.persistance.repository.JpaProviderRepository;
-import com.yes255.yes255booksusersserver.presentation.dto.request.*;
-import com.yes255.yes255booksusersserver.presentation.dto.response.FindUserResponse;
-import com.yes255.yes255booksusersserver.presentation.dto.response.LoginUserResponse;
-import com.yes255.yes255booksusersserver.presentation.dto.response.UpdateUserResponse;
-import com.yes255.yes255booksusersserver.presentation.dto.response.UserResponse;
+import com.yes255.yes255booksusersserver.presentation.dto.request.user.*;
+import com.yes255.yes255booksusersserver.presentation.dto.response.user.FindUserResponse;
+import com.yes255.yes255booksusersserver.presentation.dto.response.user.LoginUserResponse;
+import com.yes255.yes255booksusersserver.presentation.dto.response.user.UpdateUserResponse;
+import com.yes255.yes255booksusersserver.presentation.dto.response.user.UserResponse;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.context.request.RequestContextHolder;
+import org.springframework.web.context.request.ServletRequestAttributes;
 
 import java.util.List;
+import java.util.Objects;
 
+@Tag(name = "회원 API", description = "회원 관련 API 입니다.")
 @RestController
 @RequiredArgsConstructor
 public class UserController {
 
     private final UserService userService;
 
-    // 인증을 위한 회원 정보 반환
+    @Operation(summary = "인증을 위한 회원 정보 반환", description = "로그인 인증을 위한 회원 정보를 반환합니다.")
     @PostMapping("/users")
     public ResponseEntity<LoginUserResponse> findLoginUserByEmail(@RequestBody LoginUserRequest userRequest) {
         return new ResponseEntity<>(userService.findLoginUserByEmailByPassword(userRequest), HttpStatus.OK);
     }
 
-    // 회원 가입
+    @Operation(summary = "회원 가입", description = "회원 가입을 진행합니다.")
     @PostMapping("/auth/signup")
     public ResponseEntity<UserResponse> signUp(@RequestBody CreateUserRequest userRequest) {
         return new ResponseEntity<>(userService.createUser(userRequest), HttpStatus.CREATED);
     }
 
-    // 유저 로그인
+    @Operation(summary = "회원 로그인", description = "회원의 로그인을 진행합니다.")
     @PostMapping("/auth/login")
     public ResponseEntity<Boolean> login(@RequestBody LoginUserRequest userRequest) {
         return new ResponseEntity<>(userService.loginUserByEmailByPassword(userRequest), HttpStatus.OK);
     }
 
 
-    // 이메일(아이디) 찾기 (이름과 이메일이 조건에 일치하는)
+    @Operation(summary = "이메일 찾기", description = "이메일(아이디)을 찾기 위해 회원 이름과 이메일을 통해 이메일 목록을 조회합니다.")
     @PostMapping("/auth/findEmail")
     public ResponseEntity<List<FindUserResponse>> findAllByUserNameByUserPhone(@RequestBody FindEmailRequest emailRequestRequest,
                                                                                Pageable pageable) {
@@ -49,34 +54,55 @@ public class UserController {
                 , HttpStatus.OK);
     }
 
-    // 비밀번호 찾기
+    @Operation(summary = "비밀번호 찾기", description = "이메일과 회원 이름을 통해 비밀번호 찾기 인증을 진행합니다.")
     @PostMapping("/auth/findPassword")
     public ResponseEntity<Boolean> findPassword(@RequestBody FindPasswordRequest passwordRequest) {
         return new ResponseEntity<>(userService.findUserPasswordByEmailByName(passwordRequest), HttpStatus.OK);
     }
 
     // todo : 비밀번호 설정
+//    @Operation(summary = "비밀번호 설정", description = "비밀번호를 재설정합니다.")
 //    @PostMapping("/auth/setPassword")
 
 
+    @Operation(summary = "회원 조회", description = "특정 회원 정보를 조회합니다.")
+    @GetMapping("/users")
+    public ResponseEntity<UserResponse> findByUserId() {
 
-    // 회원 조회
-    @GetMapping("/users/{userId}")
-    public ResponseEntity<UserResponse> findByUserId(@PathVariable Long userId) {
+        HttpServletRequest request = ((ServletRequestAttributes) Objects.requireNonNull(RequestContextHolder.getRequestAttributes())).getRequest();
+        Long userId = (Long) request.getAttribute("userId");
+
+        if (userId == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         return new ResponseEntity<>(userService.findUserByUserId(userId), HttpStatus.OK);
     }
 
-    // 회원 수정
-    @PutMapping("/users/{userId}")
-    public ResponseEntity<UpdateUserResponse> updateUser(@PathVariable Long userId,
-                                                         @RequestBody UpdateUserRequest userRequest) {
+    @Operation(summary = "회원 수정", description = "특정 회원 정보를 수정합니다.")
+    @PutMapping("/users")
+    public ResponseEntity<UpdateUserResponse> updateUser(@RequestBody UpdateUserRequest userRequest) {
+
+        HttpServletRequest request = ((ServletRequestAttributes) Objects.requireNonNull(RequestContextHolder.getRequestAttributes())).getRequest();
+        Long userId = (Long) request.getAttribute("userId");
+
+        if (userId == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+
         return new ResponseEntity<>(userService.updateUser(userId, userRequest), HttpStatus.OK);
     }
 
-    // 회원 탈퇴
-    @DeleteMapping("/users/{userId}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long userId,
-                                                   @RequestBody DeleteUserRequest userRequest) {
+    @Operation(summary = "회원 탈퇴", description = "특정 회원이 탈퇴합니다.")
+    @DeleteMapping("/users")
+    public ResponseEntity<Void> deleteUser(@RequestBody DeleteUserRequest userRequest) {
+
+        HttpServletRequest request = ((ServletRequestAttributes) Objects.requireNonNull(RequestContextHolder.getRequestAttributes())).getRequest();
+        Long userId = (Long) request.getAttribute("userId");
+
+        if (userId == null) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
 
         userService.deleteUser(userId, userRequest);
